Refactor to namespace the forecasts import service
